chore: rename txe_service to rpc_translator
Long overdue.
